pyproject.toml
[tool.poetry]
name = "keerda-ai-backend"
version = "0.1.0"
description = ""
authors = ["RohanJnr <rohanjnr44@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
wagtail = "^5.0.2"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
